Extract Scheme-specific types into files
