+++ Sources/ProjectDescription/ArchiveAction.swift
@@ -1,0 +1,23 @@
+import Foundation
+
+public struct ArchiveAction: Equatable, Codable {
+    public let configurationName: String
+    public let revealArchiveInOrganizer: Bool
+    public let customArchiveName: String?
+    public let preActions: [ExecutionAction]
+    public let postActions: [ExecutionAction]
+
+    public init(
+        configurationName: String,
+        revealArchiveInOrganizer: Bool = true,
+        customArchiveName: String? = nil,
+        preActions: [ExecutionAction] = [],
+        postActions: [ExecutionAction] = []
+    ) {
+        self.configurationName = configurationName
+        self.revealArchiveInOrganizer = revealArchiveInOrganizer
+        self.customArchiveName = customArchiveName
+        self.preActions = preActions
+        self.postActions = postActions
+    }
+}
