+++ Sources/TuistCore/Models/Arguments.swift
@@ -1,0 +1,24 @@
+import Basic
+import Foundation
+
+public struct Arguments: Equatable {
+    // MARK: - Attributes
+
+    public let environment: [String: String]
+    public let launch: [String: Bool]
+
+    // MARK: - Init
+
+    public init(environment: [String: String] = [:],
+                launch: [String: Bool] = [:]) {
+        self.environment = environment
+        self.launch = launch
+    }
+
+    // MARK: - Equatable
+
+    public static func == (lhs: Arguments, rhs: Arguments) -> Bool {
+        lhs.environment == rhs.environment &&
+            lhs.launch == rhs.launch
+    }
+}
