+++ Sources/TuistCore/Models/ExecuteAction.swift
@@ -1,0 +1,26 @@
+import Basic
+import Foundation
+
+public struct ExecutionAction: Equatable {
+    // MARK: - Attributes
+
+    public let title: String
+    public let scriptText: String
+    public let target: TargetReference?
+
+    // MARK: - Init
+
+    public init(title: String,
+                scriptText: String,
+                target: TargetReference?) {
+        self.title = title
+        self.scriptText = scriptText
+        self.target = target
+    }
+
+    public static func == (lhs: ExecutionAction, rhs: ExecutionAction) -> Bool {
+        lhs.title == rhs.title &&
+            lhs.scriptText == rhs.scriptText &&
+            lhs.target == rhs.target
+    }
+}
