+++ Sources/TuistCore/Models/Scheme.swift
@@ -56,63 +56,3 @@ public class Scheme: Equatable {
             lhs.archiveAction == rhs.archiveAction
     }
 }
-
-public struct Arguments: Equatable {
-    // MARK: - Attributes
-
-    public let environment: [String: String]
-    public let launch: [String: Bool]
-
-    // MARK: - Init
-
-    public init(environment: [String: String] = [:],
-                launch: [String: Bool] = [:]) {
-        self.environment = environment
-        self.launch = launch
-    }
-
-    // MARK: - Equatable
-
-    public static func == (lhs: Arguments, rhs: Arguments) -> Bool {
-        lhs.environment == rhs.environment &&
-            lhs.launch == rhs.launch
-    }
-}
-
-public struct ExecutionAction: Equatable {
-    // MARK: - Attributes
-
-    public let title: String
-    public let scriptText: String
-    public let target: TargetReference?
-
-    // MARK: - Init
-
-    public init(title: String,
-                scriptText: String,
-                target: TargetReference?) {
-        self.title = title
-        self.scriptText = scriptText
-        self.target = target
-    }
-
-    public static func == (lhs: ExecutionAction, rhs: ExecutionAction) -> Bool {
-        lhs.title == rhs.title &&
-            lhs.scriptText == rhs.scriptText &&
-            lhs.target == rhs.target
-    }
-}
-
-public struct TargetReference: Hashable {
-    public var projectPath: AbsolutePath
-    public var name: String
-
-    public static func project(path: AbsolutePath, target: String) -> TargetReference {
-        .init(projectPath: path, name: target)
-    }
-
-    public init(projectPath: AbsolutePath, name: String) {
-        self.projectPath = projectPath
-        self.name = name
-    }
-}
